refactor: car 의 패키지로 Car class 이동
