+++ src/test/java/racinggame/domain/CarHistoryTest.java
@@ -1,10 +1,9 @@
 package racinggame.domain;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racinggame.domain.car.Car;
 
 class CarHistoryTest {
