+++ src/test/java/racinggame/domain/CarTest.java
@@ -1,10 +1,10 @@
 package racinggame.domain;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racinggame.domain.car.Car;
+import racinggame.domain.car.Cars;
 
 class CarTest {
